feat: add status icon
